Refactor llm_task and llm_chat, introduce coversation context to history
